+++ CHANGES.rst
@@ -9,8 +9,12 @@ New Features
 
 - Add Catalan localisation (`#966`_, welpo)
 - Make <code class="language-$lang"> for syntax highlighting (`#998`_, pkvach)
+- Add search for comments by URL in the admin interface (`#1000`_, pkvach)
 - Add CSS variables for better organization and flexibility (`#1001`_, pkvach)
 
+.. _#966: https://github.com/posativ/isso/pull/966
+.. _#998: https://github.com/isso-comments/isso/pull/998
+.. _#1000: https://github.com/isso-comments/isso/pull/1000
 .. _#966: https://github.com/posativ/isso/pull/966
 .. _#998: https://github.com/isso-comments/isso/pull/998
 .. _#1001: https://github.com/isso-comments/isso/pull/1001